Refactor batch worker internals
